refactor(acceptance): repoint suite paths after moving under suites/
The mcp and downstream-mtls suites are one level deeper now, so: walk two parents (not one) to reach the shared internal/acceptance root for the Keycloak realm mount and the parent OpenSSL cert scripts; bump the mcp -> browser/fixtures/users.js imports one extra hop; and point the root Makefile deps-/test- delegates at suites/mcp and suites/downstream-mtls. No shared code yet; each suite still stands alone.
