+++ internal/acceptance/suites/downstream-mtls/setup/constants.ts
@@ -24,7 +24,7 @@ export const MTLS_CERTS_DIR = path.join(CERTS_DIR, "mtls");
 /** CA that signs the server TLS leaf; used to verify Pomerium's certificate. */
 export const SERVER_CA_FILE = path.join(CERTS_DIR, "ca.crt");
 
-/** Test user from the shared acceptance realm (../keycloak). The mTLS client
+/** Test user from the shared acceptance realm (../../keycloak). The mTLS client
  * certificates are issued to the same identity (SAN email alice@company.com). */
 export const TEST_USER = {
   email: "alice@company.com",
